+++ opennmt/tests/config_test.py
@@ -1,4 +1,5 @@
 import os
+import filecmp
 import yaml
 
 import tensorflow as tf
@@ -26,10 +27,14 @@ class ConfigTest(tf.test.TestCase):
         {"model_dir": "bar", "train": {"batch_size": 64, "steps": 42}},
         loaded_config)
 
-  def testLoadModelModule(self):
-    model_path = os.path.join(self.get_temp_dir(), "model.py")
+  def _writeCustomModel(self, filename="test_model.py", return_value=42):
+    model_path = os.path.join(self.get_temp_dir(), filename)
     with open(model_path, mode="wb") as model_file:
-      model_file.write(b"model = lambda: 42")
+      model_file.write(b"model = lambda: %d" % return_value)
+    return model_path
+
+  def testLoadModelModule(self):
+    model_path = self._writeCustomModel()
     model_module = config.load_model_module(model_path)
     model = model_module.model()
     self.assertEqual(42, model)
@@ -43,9 +48,31 @@ class ConfigTest(tf.test.TestCase):
     model_name = "NMTSmall"
     model_dir = self.get_temp_dir()
     model = config.load_model(model_dir, model_name=model_name)
+    self.assertTrue(os.path.exists(os.path.join(model_dir, "model_description.py")))
     self.assertIsInstance(model, Model)
     model = config.load_model(model_dir)
     self.assertIsInstance(model, Model)
+
+  def testLoadModelFile(self):
+    model_file = self._writeCustomModel()
+    model_dir = self.get_temp_dir()
+    model = config.load_model(model_dir, model_file=model_file)
+    saved_description_path = os.path.join(model_dir, "model_description.py")
+    self.assertTrue(os.path.exists(saved_description_path))
+    self.assertTrue(filecmp.cmp(model_file, saved_description_path))
+    self.assertEqual(model, 42)
+    model = config.load_model(model_dir)
+    self.assertEqual(model, 42)
+
+  def testLoadModelFileOverride(self):
+    model_dir = self.get_temp_dir()
+    saved_description_path = os.path.join(model_dir, "model_description.py")
+    model_file = self._writeCustomModel(filename="test_model1.py", return_value=1)
+    model = config.load_model(model_dir, model_file=model_file)
+    self.assertTrue(filecmp.cmp(model_file, saved_description_path))
+    model_file = self._writeCustomModel(filename="test_model2.py", return_value=2)
+    model = config.load_model(model_dir, model_file=model_file)
+    self.assertTrue(filecmp.cmp(model_file, saved_description_path))
 
   def testLoadModelInvalidArguments(self):
     with self.assertRaises(ValueError):